I changed around winbolo's tank hit code to make it so winbolo gently pushes the tank out of the blast radius, I did this by making it move the tank 16 world coordinates 8 times for a total of 128 world coordinates. This is similiar to macbolo's tank slide. Thanks goes out to sticks for this one, his work on tankslide helped out alot.
I decreased the swamp life by one so that swamp dies in 4 shots, for some reason the variable has to be set to 3 to allow it to die in 4 shots. Thanks goes to nemo for noticing the swamp hits. winbolo used to take 5 shots to kill a swamp square, macbolo is 4 shots. This is fixed now